requirements.txt
discord.py>=2.3
openai>=1.0
groq>=0.4
aiosqlite>=0.19
python-dotenv>=1.0
